requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
jieba==0.42.1
lxml==4.6.2
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.5
pdfminer3k==1.3.4
ply==3.11
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.5
six==1.15.0
